move osm parsing to io module. rename feature gates
