--- h3ron-graph/src/formats/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "osmpbfreader")]
-pub mod osm;
