+++ h3ron-graph/src/io/mod.rs
@@ -1,2 +1,4 @@
-#[cfg(feature = "use-gdal")]
+#[cfg(feature = "io_gdal")]
 pub mod gdal;
+#[cfg(feature = "io_osm")]
+pub mod osm;
